extracted add category to provider
