+++ client/src/components/Blog/Categories/Categories.js
@@ -5,20 +5,21 @@ import CategoriesList from "./CategoriesList";
 import { Container, Button } from "semantic-ui-react";
 import {Link, } from 'react-router-dom'
 import styled from "styled-components";
+import {PostCatConsumer, } from "../../../providers/PostCatProvider"
 
 class Categories extends React.Component {
-  state = { categories: [] };
+  state = { categories: this.props.postcat.categories };
 
-  componentDidMount() {
-    axios
-      .get("/api/categories")
-      .then(res => {
-        this.setState({ categories: res.data });
-      })
-      .catch(err => {
-        console.log(err);
-      });
-  }
+  // componentDidMount() {
+  //   axios
+  //     .get("/api/categories")
+  //     .then(res => {
+  //       this.setState({ categories: res.data });
+  //     })
+  //     .catch(err => {
+  //       console.log(err);
+  //     });
+  // }
 
   componentDidUpdate(prevProps, prevState) {
     if (this.state.categories !== prevState.categories) {
@@ -31,36 +32,36 @@ class Categories extends React.Component {
     }
   }
 
-  addCategory = label => {
-    axios.post("/api/categories", { label }).then(res => {
-      const { categories } = this.state;
-      this.setState({ categories: [...categories, res.data] });
-    });
-  };
+  // addCategory = label => {
+  //   axios.post("/api/categories", { label }).then(res => {
+  //     const { categories } = this.state;
+  //     this.setState({ categories: [...categories, res.data] });
+  //   });
+  // };
 
-  editCategory = category => {
-    axios.put(`/api/categories/${category.id}`, category).then(res => {
-      const categories = this.state.categories.map(cat => {
-        if (cat.id === category.id) return res.data;
-        return cat;
-      });
-      this.setState({ categories });
-    });
-  };
+  // editCategory = category => {
+  //   axios.put(`/api/categories/${category.id}`, category).then(res => {
+  //     const categories = this.state.categories.map(cat => {
+  //       if (cat.id === category.id) return res.data;
+  //       return cat;
+  //     });
+  //     this.setState({ categories });
+  //   });
+  // };
 
-  deleteCategory = id => {
-    axios.delete(`/api/categories/${id}`).then(res => {
-      const { categories } = this.state;
-      this.setState({ categories: categories.filter(cat => cat.id !== id) });
-    });
-  };
+  // deleteCategory = id => {
+  //   axios.delete(`/api/categories/${id}`).then(res => {
+  //     const { categories } = this.state;
+  //     this.setState({ categories: categories.filter(cat => cat.id !== id) });
+  //   });
+  // };
  
   render() {
     return (
       <StyledDiv>
         <StyledContainer>
           <br />
-          <CategoryForm addCategory={this.addCategory} />
+          <CategoryForm  />
           <Link to="/blog">
             <Button floated="right">Go Back</Button>
           </Link>
@@ -68,8 +69,8 @@ class Categories extends React.Component {
           <br />
           <CategoriesList
             categories={this.state.categories}
-            editCategory={this.editCategory}
-            deleteCategory={this.deleteCategory}
+            editCategory={this.props.postcat.editCategory}
+            deleteCategory={this.props.postcat.deleteCategory}
           />
         </StyledContainer>
       </StyledDiv>
@@ -85,4 +86,15 @@ const StyledContainer = styled(Container)`
   padding: 0em 10em !important;
 `;  
 
-export default Categories;
+export default class ConnectedCategories extends React.Component{
+  render(){
+    return(
+      <PostCatConsumer>
+        {postcat => (
+          <Categories {...this.props} postcat={postcat}/>
+        )}
+      </PostCatConsumer>
+    )
+  }
+
+}
